fix the tests

+++ src/cascadia/TerminalControl/ControlAppearance.h
@@ -40,9 +40,9 @@ namespace winrt::Microsoft::Terminal::Control::implementation
             CONTROL_APPEARANCE_SETTINGS(COPY_SETTING)
 #undef COPY_SETTING
 
-            for (int32_t i = 0; i < _ColorTable.size(); i++)
+            for (size_t i = 0; i < _ColorTable.size(); i++)
             {
-                _ColorTable[i] = appearance.GetColorTableEntry(i);
+                _ColorTable[i] = appearance.GetColorTableEntry(static_cast<int32_t>(i));
             }
         }
     };
